cloud_function/requirements.txt
google-cloud-storage==2.14.0
requests==2.31.0
functions-framework==3.5.0